New Scene Sliders and Levers
Summary:
# Context
More use cases on how to use the ISDK features Sliders and Levers. This scene goes through the proper ISDK way to use hand grab poses, transform constraints, and snap sliders.

# Product decisions
1. Levers
2. Sliders
3. Buttons
4. Scene for just the models. Used just incase users want something barebones.
5. Scene on how these models are used ISDK way

{F1974552145}

# Scripts
1. BasicToggle.cs (Buttons that slide left and right)
2. StepsMove.cs (Slider snapping at different levels)
3. ToggleMove.cs (Slider snapping on and off)

# Experimental WIP
There are a few different experimental features inside of the project that are partly finished or incomplete use with caution or we should delete them since they are not finished and or buggy.

Example: levers with buttons works but the thumb has to be in the perfect spot to hover and interact while grabbing the lever.
 {F1974772113}

Reviewed By: SaadKhan64

Differential Revision: D68607054

fbshipit-source-id: 25917a7db3c6b73437f4fcce617ba433e0650adf