Telepath Climbing and WalkingStick GitHub samples
Summary:
This diff includes many new Locomotion Interactions as explored during the 2025H1 in https://docs.google.com/document/d/1W1W4Aon5NwI6ki1OL1unm8ozwjaOmTbKZA5s5Qyi4DU/edit?tab=t.0#heading=h.79hslkb1zcrl

Test the APK here:  https://pxl.cl/7RW5Z

And here is a full video:  https://pxl.cl/7RWjZ

# Telepath

Telepath Locomotion transforms Absolute LocomotionEvents into a series of Velocity LocomotionEvents to drive the character to the desired destination. It requires the following pieces:

### TelepathActionsBroadcaster

This optional piece decorates the associated TeleportInteractors with extra information if the Telepath should be Halted or including Hover events to help visualise the Telepath during Hover (remember that TeleportInteractor only emits events during Selection).

An alternative would have been having the Locomotor and Visuals all suscribe to all TeleportInteractors manually. Which is also viable, but using this decorator helps put everything in the LocomotionEvent pipeline

### TelepathVisual

Only one is truly needed, and it receives the Hover events from TelepathActionsBroadcaster as well as normal Teleport Selections to draw the path using DotGridUnlit.shader.  This shader has been improved to support VR and be performant for rendering the dots.

### TelepathLocomotionBroadcaster

This is the actual telepath "Locomotor". But it does not directly move the character, instead it receives incoming Teleport events and transforms them into a series of Velocity events.
Because of this it needs to be placed right before the FirstPersonLocomotor (or any other Velocity consumer).

// I am strongly considering renaming this one to TelepathLocomotor.

# Climbing

The Climbing system leverages the Grabbing system but instead of consuming PointerEvents to move an object with a Transformer in the Interactable side. It converts these PointerEvents into LocomotionEvents (decorated with climbing information) that are redirected to the Interactor (GrabInteractor) side so it can be wired to the actual locomotor.

### Climbable

The climbable is a PointableElement that replaces the Grabbable. Attach it to an object in the world and have its GrabInteractables point to it. It will then transform incoming PointerEvents into LocomotionEvents.

### ClimbingLocomotionBroadcaster

It decorates an Interactor (typically a GrabInteractor) to receive the LocomotionEvents from the Climbable and put them in the LocomotionEventPipeline to be consumed by the Locomotor.

### ClimbingLocomotor

This Locomotor receives all the Climbing LocomotionEvents and transforms them into Relative/VelocityEvents. It is important to have only a central piece for Climbing as in many cases you will want to disregard certain Hands, calculate the average velocity, etc.

Originally this Locomotor was affecting the CharacterController directly, but then I removed all dependencies and made it output proper LocomotionEvents that can be consumed by the FirstPersonLocomotor (or other custom implementations)

# WalkingStick

This interaction allows using walking-sticks with hands to push yourself around.

### WalkingStick.cs

This piece stays in each Hand or COntroller and is merely used as a proxy to indicate to the Locomotor if the hand is active and help drive the visual.

### WalkingStickVisual

It draws the WalkingStick  in each hand or controller

### WalkingStickLocomotor

This is the main piece. Unlike other interactions, WalkingStick really needs to be a single piece that listens to all individual walking sticks to work together.  The code is very UX driven and it contains a ton of constants. I decided on purpose not to expose them as fields as they are not trivial to change and in general modifying them will just yield bad results

# Other considerations

### Grounded modifications

I greatly simplified the Grounded checks for CharacterController. Removing a ton of redundant code while making it more performant and reliable. I added some Tests for this.

### OpenXRValidate

The SceneBuildOpenXRValidate file had to be removed as it prevented builds with HandGrabs created purely in OpenXR mode, as well as other new offsets

### Bigger scene

The LocomotionExamples scene has grown a lot to accommodate the new interactions and the setting sin this particular scene have been greatly expanded.

### Navigation Dependency

We always had this dependency, but now a test is actively complaining about it because it can find it in our LocomotionExamples scene. We discussed to add it to the package dependencies since it is a very basic package that comes pre-installed in unity.

Reviewed By: SaadKhan64

Differential Revision: D80816504

fbshipit-source-id: 5ce0f280751e1526c57077d9a1789a1a8a765cf7